+++ src/astUtils/visitors.spec.ts
@@ -4,7 +4,7 @@ import { CancellationTokenSource, Range } from 'vscode-languageserver';
 import { expect } from 'chai';
 import * as sinon from 'sinon';
 import { Program } from '../Program';
-import { BrsFile } from '../files/BrsFile';
+import type { BrsFile } from '../files/BrsFile';
 import type { Statement } from '../parser/Statement';
 import { PrintStatement, Block, ReturnStatement } from '../parser/Statement';
 import type { Expression } from '../parser/Expression';
@@ -17,7 +17,6 @@ import { createStackedVisitor } from './stackedVisitor';
 describe('astUtils visitors', () => {
     const rootDir = process.cwd();
     let program: Program;
-    let file: BrsFile;
 
     const PRINTS_SRC = `
         sub Main()
@@ -76,7 +75,6 @@ describe('astUtils visitors', () => {
 
     beforeEach(() => {
         program = new Program({ rootDir: rootDir });
-        file = new BrsFile('abs.bs', 'rel.bs', program);
     });
     afterEach(() => {
         program.dispose();
@@ -102,9 +100,9 @@ describe('astUtils visitors', () => {
             const walker = functionsWalker(visitor);
             program.plugins.add({
                 name: 'walker',
-                afterFileParse: () => walker(file)
+                afterFileParse: file => walker(file as BrsFile)
             });
-            file.parse(PRINTS_SRC);
+            program.addOrReplaceFile('source/main.brs', PRINTS_SRC);
             expect(actual).to.deep.equal([
                 'Block:0',                // Main sub body
                 'PrintStatement:1',       // print 1
@@ -139,9 +137,9 @@ describe('astUtils visitors', () => {
             const walker = functionsWalker(s => actual.push(s.constructor.name), cancel.token);
             program.plugins.add({
                 name: 'walker',
-                afterFileParse: () => walker(file)
+                afterFileParse: file => walker(file as BrsFile)
             });
-            file.parse(PRINTS_SRC);
+            program.addOrReplaceFile('source/main.brs', PRINTS_SRC);
             expect(actual).to.deep.equal([
                 'Block',                // Main sub body
                 'PrintStatement',       // print 1
@@ -184,9 +182,9 @@ describe('astUtils visitors', () => {
             }, cancel.token);
             program.plugins.add({
                 name: 'walker',
-                afterFileParse: () => walker(file)
+                afterFileParse: file => walker(file as BrsFile)
             });
-            file.parse(PRINTS_SRC);
+            program.addOrReplaceFile('source/main.brs', PRINTS_SRC);
             expect(actual).to.deep.equal([
                 'Block',                // Main sub body
                 'PrintStatement',       // print 1
@@ -263,10 +261,10 @@ describe('astUtils visitors', () => {
             });
             program.plugins.add({
                 name: 'walker',
-                afterFileParse: () => walker(file)
+                afterFileParse: (file) => walker(file as BrsFile)
             });
 
-            file.parse(EXPRESSIONS_SRC);
+            program.addOrReplaceFile('source/main.brs', EXPRESSIONS_SRC);
             expect(actual).to.deep.equal([
                 //The comment statement is weird because it can't be both a statement and expression, but is treated that way. Just ignore it for now until we refactor comments.
                 //'CommentStatement:1:CommentStatement',          // '<comment>