+++ src/files/BrsFile.ts
@@ -268,28 +268,14 @@ export class BrsFile {
                 ...this._parser.diagnostics as BsDiagnostic[]
             );
 
-            //notify AST ready
-            this.program.plugins.emit('afterFileParse', this);
-
             //extract all callables from this file
             this.findCallables();
 
             //find all places where a sub/function is being called
             this.findFunctionCalls();
 
-            //emit an event before starting to validate this file
-            this.program.plugins.emit('beforeFileValidate', {
-                file: this,
-                program: this.program
-            });
-
-            //emit an event to allow plugins to contribute to the file validation process
-            this.program.plugins.emit('onFileValidate', {
-                file: this,
-                program: this.program
-            });
-
-            this.findAndValidateImportAndImportStatements();
+            //register all import statements for use in the rest of the program
+            this.registerImports();
 
             //attach this file to every diagnostic
             for (let diagnostic of this.diagnostics) {
@@ -305,9 +291,29 @@ export class BrsFile {
         }
     }
 
-    public findAndValidateImportAndImportStatements() {
-        let topOfFileIncludeStatements = [] as Array<LibraryStatement | ImportStatement>;
+    private registerImports() {
+        for (const statement of this.parser?.references?.importStatements ?? []) {
+            //register import statements
+            if (isImportStatement(statement) && statement.filePathToken) {
+                this.ownScriptImports.push({
+                    filePathRange: statement.filePathToken.range,
+                    pkgPath: util.getPkgPathFromTarget(this.pkgPath, statement.filePath),
+                    sourceFile: this,
+                    text: statement.filePathToken?.text
+                });
+            }
+        }
+    }
 
+    public validate() {
+        //only validate the file if it was actually parsed (skip files containing typedefs)
+        if (!this.hasTypedef) {
+            this.validateImportStatements();
+        }
+    }
+
+    private validateImportStatements() {
+        let topOfFileIncludeStatements = [] as Array<LibraryStatement | ImportStatement>;
         for (let stmt of this.ast.statements) {
             //skip comments
             if (isCommentStatement(stmt)) {
@@ -327,16 +333,6 @@ export class BrsFile {
             ...this._parser.references.importStatements
         ];
         for (let result of statements) {
-            //register import statements
-            if (isImportStatement(result) && result.filePathToken) {
-                this.ownScriptImports.push({
-                    filePathRange: result.filePathToken.range,
-                    pkgPath: util.getPkgPathFromTarget(this.pkgPath, result.filePath),
-                    sourceFile: this,
-                    text: result.filePathToken?.text
-                });
-            }
-
             //if this statement is not one of the top-of-file statements,
             //then add a diagnostic explaining that it is invalid
             if (!topOfFileIncludeStatements.includes(result)) {
@@ -1487,6 +1483,7 @@ export class BrsFile {
     }
 
     public getHover(position: Position): Hover {
+        const fence = (code: string) => util.mdFence(code, 'brightscript');
         //get the token at the position
         let token = this.getTokenAt(position);
 
@@ -1523,7 +1520,7 @@ export class BrsFile {
                         return {
                             range: token.range,
                             //append the variable name to the front for scope
-                            contents: typeText
+                            contents: fence(typeText)
                         };
                     }
                 }
@@ -1531,7 +1528,7 @@ export class BrsFile {
                     if (labelStatement.name.toLocaleLowerCase() === lowerTokenText) {
                         return {
                             range: token.range,
-                            contents: `${labelStatement.name}: label`
+                            contents: fence(`${labelStatement.name}: label`)
                         };
                     }
                 }
@@ -1546,11 +1543,39 @@ export class BrsFile {
                 if (callable) {
                     return {
                         range: token.range,
-                        contents: callable.type.toString()
+                        contents: this.getCallableDocumentation(callable)
                     };
                 }
             }
         }
+    }
+
+    /**
+     * Build a hover documentation for a callable.
+     */
+    private getCallableDocumentation(callable: Callable) {
+        const comments = [] as Token[];
+        const tokens = callable.file.parser.tokens as Token[];
+        const idx = tokens.indexOf(callable.functionStatement?.func.functionType);
+        for (let i = idx - 1; i >= 0; i--) {
+            const token = tokens[i];
+            //skip whitespace and newline chars
+            if (token.kind === TokenKind.Comment) {
+                comments.push(token);
+            } else if (token.kind === TokenKind.Newline || token.kind === TokenKind.Whitespace) {
+                //skip these tokens
+                continue;
+
+                //any other token means there are no more comments
+            } else {
+                break;
+            }
+        }
+        let result = util.mdFence(callable.type.toString(), 'brightscript');
+        if (comments.length > 0) {
+            result += '\n***\n' + comments.reverse().map(x => x.text.replace(/^('|rem)/i, '')).join('\n');
+        }
+        return result;
     }
 
     public getSignatureHelpForNamespaceMethods(callableName: string, dottedGetText: string, scope: Scope): { key: string; signature: SignatureInformation }[] {